hot fix

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned long n = 0;
+	size_t n = 0;
 
 	while (1)
 	{
